+++ src/main/java/com/sparta/finalproject/FinalProjectApplication.java
@@ -6,9 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class FinalProjectApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(FinalProjectApplication.class, args);
-        System.out.println("테스트용");
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(FinalProjectApplication.class, args);
+	}
 
 }
