--- src/main/java/com/sparta/finalproject/common/config/SecurityConfig.java
@@ -1,60 +1,0 @@
-package com.sparta.finalproject.common.config;
-
-import com.sparta.finalproject.common.jwt.JwtAuthFilter;
-import com.sparta.finalproject.common.jwt.JwtUtil;
-import com.sparta.finalproject.common.security.CustomAccessDeniedHandler;
-import com.sparta.finalproject.common.security.CustomAuthenticationEntryPoint;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-@EnableMethodSecurity(prePostEnabled = true)
-public class SecurityConfig {
-
-  private final JwtUtil jwtUtil;
-
-  @Bean
-  public PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder();
-  }
-
-  @Bean
-  public WebSecurityCustomizer webSecurityCustomizer() {
-    return (web -> web.ignoring()
-        .requestMatchers(PathRequest.toH2Console()));
-  }
-
-  @Bean
-  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-    http
-        .httpBasic().disable()
-        .csrf().disable()
-        .formLogin().disable()
-        .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        .and()
-        .authorizeHttpRequests()
-        .antMatchers("/api/auth/**").permitAll()
-        .anyRequest().authenticated()
-        .and()
-        .exceptionHandling().authenticationEntryPoint(new CustomAuthenticationEntryPoint())
-        .and()
-        .exceptionHandling().accessDeniedHandler(new CustomAccessDeniedHandler())
-        .and()
-        .addFilterBefore(new JwtAuthFilter(jwtUtil), UsernamePasswordAuthenticationFilter.class);
-
-    return http.build();
-  }
-}
