--- src/main/java/com/sparta/finalproject/common/jwt/JwtUtil.java
@@ -1,121 +1,0 @@
-package com.sparta.finalproject.common.jwt;
-
-import com.sparta.finalproject.user.entity.UserRole;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.UnsupportedJwtException;
-import io.jsonwebtoken.security.Keys;
-import io.jsonwebtoken.security.SecurityException;
-import java.security.Key;
-import java.util.Base64;
-import java.util.Date;
-import javax.annotation.PostConstruct;
-import javax.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class JwtUtil {
-
-  private final UserDetailsService userDetailsService;
-  public static final String AUTHORIZATION_HEADER = "Authorization";
-  public static final String REFRESH_HEADER = "refresh_token";
-  public static final String AUTHORIZATION_KEY = "auth";
-  private static final String BEARER_PREFIX = "Bearer ";
-  private static final long TOKEN_TIME = 60 * 60 * 1000L;
-  private static final long REFRESH_TOKEN_VALID_TIME = 60 * 60 * 24 * 7 * 1000L; // 7 일
-
-  @Value("${jwt.secret.key}")
-  private String secretKey;
-  private Key key;
-  private final SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS256;
-
-  @PostConstruct
-  public void init() {
-    byte[] bytes = Base64.getDecoder().decode(secretKey);
-    key = Keys.hmacShaKeyFor(bytes);
-  }
-
-  // header 토큰 가져오기
-  public String resolveToken(HttpServletRequest request) {
-    String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
-    if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(BEARER_PREFIX)) {
-      return bearerToken.substring(7);
-    }
-    return null;
-
-  }
-
-  // 토큰 생성
-  public String createToken(String username, UserRole userRole) {
-    Date date = new Date();
-
-    return BEARER_PREFIX +
-        Jwts.builder()
-            .setSubject(username)
-            .claim(AUTHORIZATION_KEY, userRole)
-            .setExpiration(new Date(date.getTime() + TOKEN_TIME))
-            .setIssuedAt(date)
-            .signWith(key, signatureAlgorithm)
-            .compact();
-  }
-
-  public String createRefreshToken() {
-    Date now = new Date();
-
-    return Jwts.builder()
-        .setIssuedAt(now)
-        .setExpiration(new Date(now.getTime() + REFRESH_TOKEN_VALID_TIME))
-        .signWith(key, signatureAlgorithm)
-        .compact();
-  }
-
-  // 토큰 검증
-  public boolean validateToken(String token) {
-    try {
-      Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-      return true;
-    } catch (SecurityException | MalformedJwtException e) {
-      log.info("Invalid JWT signature, 유효하지 않는 JWT 서명 입니다.");
-    } catch (ExpiredJwtException e) {
-      log.info("Expired JWT token, 만료된 JWT token 입니다.");
-    } catch (UnsupportedJwtException e) {
-      log.info("Unsupported JWT token, 지원되지 않는 JWT 토큰 입니다.");
-    } catch (IllegalArgumentException e) {
-      log.info("JWT claims is empty, 잘못된 JWT 토큰 입니다.");
-    }
-    return false;
-  }
-
-  public Claims getUserInfoFromToken(String token) {
-    return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).getBody();
-  }
-
-  public Authentication createAuthentication(String username) {
-    UserDetails userDetails = userDetailsService.loadUserByUsername(username);
-    return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-  }
-
-  public boolean validateTokenExceptExpiration(String token) {
-    try {
-      Jws<Claims> claims = Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-      return !claims.getBody().getExpiration().before(new Date());
-    } catch (Exception e) {
-      return false;
-    }
-  }
-
-}
