--- src/main/java/com/sparta/finalproject/common/security/CustomAccessDeniedHandler.java
@@ -1,29 +1,0 @@
-package com.sparta.finalproject.common.security;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.json.JSONObject;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-
-@Slf4j
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
-
-  @Override
-  public void handle(HttpServletRequest request, HttpServletResponse response,
-      AccessDeniedException accessDeniedException) throws IOException, ServletException {
-    log.error("권한 없는 사용자의 접근");
-
-    response.setContentType("application/json;charset=UTF-8");
-    response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-
-    JSONObject responseJson = new JSONObject();
-    responseJson.put("message", "권한 없는 사용자의 접근입니다.");
-    responseJson.put("code", HttpServletResponse.SC_FORBIDDEN);
-
-    response.getWriter().print(responseJson);
-  }
-}
