--- src/main/java/com/sparta/finalproject/common/security/CustomAuthenticationEntryPoint.java
@@ -1,29 +1,0 @@
-package com.sparta.finalproject.common.security;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.json.JSONObject;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-
-@Slf4j
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-  @Override
-  public void commence(HttpServletRequest request, HttpServletResponse response,
-      AuthenticationException authException) throws IOException, ServletException {
-    log.error("가입되지 않은 사용자 접근");
-
-    response.setContentType("application/json;charset=UTF-8");
-    response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-    JSONObject responseJson = new JSONObject();
-    responseJson.put("message", "가입되지 않은 사용자의 접근입니다.");
-    responseJson.put("code", 401);
-
-    response.getWriter().print(responseJson);
-  }
-}
