--- src/main/java/com/sparta/finalproject/common/security/UserDetailsImpl.java
@@ -1,66 +1,0 @@
-package com.sparta.finalproject.common.security;
-
-import com.sparta.finalproject.user.entity.User;
-import com.sparta.finalproject.user.entity.UserRole;
-import java.util.ArrayList;
-import java.util.Collection;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class UserDetailsImpl implements UserDetails {
-
-  private final User user;
-  private final String userId;
-
-  public UserDetailsImpl(User user, String userId) {
-    this.user = user;
-    this.userId = userId;
-  }
-
-  public User getUser() {
-    return user;
-  }
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    UserRole role = user.getRole();
-    String authority = role.getAuthority();
-
-    SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
-    Collection<GrantedAuthority> authorities = new ArrayList<>();
-    authorities.add(simpleGrantedAuthority);
-
-    return authorities;
-  }
-
-  @Override
-  public String getPassword() {
-    return null;
-  }
-
-  @Override
-  public String getUsername() {
-    return userId;
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return false;
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return false;
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return false;
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return false;
-  }
-}
