--- src/main/java/com/sparta/finalproject/user/entity/UserRole.java
@@ -1,24 +1,0 @@
-package com.sparta.finalproject.user.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum UserRole {
-    USER(Authority.USER),  // 사용자 권한
-    ADMIN(Authority.ADMIN);  // 관리자 권한
-
-    private final String authority;
-
-    UserRole(String authority) {
-        this.authority = authority;
-    }
-
-    public String getAuthority() {
-        return this.authority;
-    }
-
-    public static class Authority {
-        public static final String USER = "ROLE_USER";
-        public static final String ADMIN = "ROLE_ADMIN";
-    }
-}
