--- src/main/java/com/sparta/finalproject/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.sparta.finalproject.user.repository;
-
-import com.sparta.finalproject.user.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByUserId(String userId);
-  
-}
